// apps/stroltm/ui/src/api/generated/api.ts
export interface ApiuResultSuccess {
  /**
   *
   * @type {string}
   * @memberof ApiuResultSuccess
   */
  data?: string;
}
// ...
export interface FormattedStats {
  /**
   *
   * @type {number}
   * @memberof FormattedStats
   */
  snapshotsCount?: number;
  /**
   *
   * @type {number}
   * @memberof FormattedStats
   */
  totalFileCount?: number;
  /**
   *
   * @type {number}
   * @memberof FormattedStats
   */
  totalSize?: number;
  /**
   *
   * @type {string}
   * @memberof FormattedStats
   */
  totalSizeFormatted?: string;
}
// ...
export interface ManagerTaskItem {
  /**
   *
   * @type {boolean}
   * @memberof ManagerTaskItem
   */
  isRunning?: boolean;
  /**
   *
   * @type {string}
   * @memberof ManagerTaskItem
   */
  lastEndedAt?: string;
  /**
   *
   * @type {OperationType}
   * @memberof ManagerTaskItem
   */
  operation?: OperationType;
  /**
   *
   * @type {string}
   * @memberof ManagerTaskItem
   */
  serviceName?: string;
  /**
   *
   * @type {string}
   * @memberof ManagerTaskItem
   */
  startedAt?: string;
  /**
   *
   * @type {string}
   * @memberof ManagerTaskItem
   */
  taskName?: string;
  /**
   *
   * @type {TriggerType}
   * @memberof ManagerTaskItem
   */
  trigger?: TriggerType;
}
/**
 *
 * @export
 * @interface ManagerhBackupAllResponse
 */
export interface ManagerhBackupAllResponse {
  /**
   *
   * @type {Array<ManagerhBackupAllStatusItem>}
   * @memberof ManagerhBackupAllResponse
   */
  errorStarted?: Array<ManagerhBackupAllStatusItem>;
  /**
   *
   * @type {Array<ManagerhBackupAllStatusItem>}
   * @memberof ManagerhBackupAllResponse
   */
  successStarted?: Array<ManagerhBackupAllStatusItem>;
}
/**
 *
 * @export
 * @interface ManagerhBackupAllStatusItem
 */
export interface ManagerhBackupAllStatusItem {
  /**
   *
   * @type {string}
   * @memberof ManagerhBackupAllStatusItem
   */
  instanceName?: string;
  /**
   *
   * @type {string}
   * @memberof ManagerhBackupAllStatusItem
   */
  proxyName?: string;
  /**
   *
   * @type {string}
   * @memberof ManagerhBackupAllStatusItem
   */
  serviceName?: string;
  /**
   *
   * @type {string}
   * @memberof ManagerhBackupAllStatusItem
   */
  taskName?: string;
}
/**
 *
 * @export
 * @enum {string}
 */

export const OperationType = {
  OperationTypeBACKUP: "BACKUP",
  OperationTypePRUNE: "PRUNE",
  OperationTypeRESTORE: "RESTORE",
  OperationTypeSNAPSHOTS: "SNAPSHOTS",
  OperationTypeSTATS: "STATS",
} as const;

export type OperationType = (typeof OperationType)[keyof typeof OperationType];

/**
 *
 * @export
 * @interface PublicGetPingResponse
 */
export interface PublicGetPingResponse {
  /**
   *
   * @type {string}
   * @memberof PublicGetPingResponse
   */
  data?: string;
}
/**
 *
 * @export
 * @interface ServicesGetPruneResult
 */
export interface ServicesGetPruneResult {
  /**
   *
   * @type {Array<Snapshot>}
   * @memberof ServicesGetPruneResult
   */
  data?: Array<Snapshot>;
}
/**
 *
 * @export
 * @interface ServicesGetSnapshotsResult
 */
export interface ServicesGetSnapshotsResult {
  /**
   *
   * @type {Array<Snapshot>}
   * @memberof ServicesGetSnapshotsResult
   */
  items?: Array<Snapshot>;
}
/**
 *
 * @export
 * @interface ServicesGetStatsResult
 */
export interface ServicesGetStatsResult {
  /**
   *
   * @type {FormattedStats}
   * @memberof ServicesGetStatsResult
   */
  data?: FormattedStats;
}
/**
 *
 * @export
 * @interface Snapshot
 */
export interface Snapshot {
  /**
   *
   * @type {string}
   * @memberof Snapshot
   */
  id?: string;
  /**
   *
   * @type {Array<string>}
   * @memberof Snapshot
   */
  paths?: Array<string>;
  /**
   *
   * @type {string}
   * @memberof Snapshot
   */
  shortId?: string;
  /**
   *
   * @type {Array<string>}
   * @memberof Snapshot
   */
  tags?: Array<string>;
  /**
   *
   * @type {string}
   * @memberof Snapshot
   */
  time?: string;
}
/**
 *
 * @export
 * @enum {string}
 */

export const TriggerType = {
  TriggerTypeSCHEDULE: "SCHEDULE",
  TriggerTypeAPI: "API",
  TriggerTypeMANUAL: "MANUAL",
} as const;

export type TriggerType = (typeof TriggerType)[keyof typeof TriggerType];